requirements.txt
# SWARM Research OS — Python dependencies
pyyaml>=6.0
jsonschema>=4.0
